Update README.md
Credit card fraud is a major concern for banks and financial institutions. Fraudsters use various techniques to steal credit card information and make unauthorized transactions. In this project, we will explore a dataset containing credit card transactions and build models to predict fraudulent transactions.

The purpose of this is to determine which crdit card is more likely to be fraudulent which is very important in financial institutions.